Initial commit of Origin `oct` Jenkins jobs
The set of Jenkins jobs that are run on a pull request to Origin that
use the `origin-ci-tool` are generated using the tools in this commit.
To change the set of jobs that are included in a full Origin test run,
change `children/test_pull_requests_origin_future.txt`. To change the
commands that are run in any specific test, change the corresponding
file under `commands/`.

Re-generate the set of job XMLs with `generate.sh`.

Push the new job XMLs to the Jenkins server by running `push-update.sh`
with `$USERNAME` and `$PASSWORD` set.

Verify that any new commit you make is checking in the accompanying
generated changes with `verify.sh`.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>