Touch latest installation when it's used
The logic to prune installation directories on the Jenkins master looks
at the last modified date; when the installation is more than a day old
and is not the installation pointed to with the `/latest` symlink, the
job removes it. This algorithm can crash jobs, however, if the latest
installation is very old and a new one is made, then the prune job runs.
In this case, a job could have been started before the upgrade and would
be using the old version, but the prune job would not know this. If we
touch the installation directory we are using in the job when we use it,
we keep it "new" and the prune job will only remove installations that
have not been used for a day, which is reasonable as no job should run
that long.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>