Added artifact generator, systemd journal fetcher
Often more specific artifacts are necessary than what already exists on
the remote host. In these cases, it is necessary to generate these
artifacts by actions on the host, then grab the output. A special case
of this is the sytemd journal for a unit, so it is broken out for
convenience.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>