Add `$ETCD_DATA_DIR` to system environment
We need `etcd` to store it's data in a tmpfs to ensure that the weird
semantics around IO on EC2 EBS don't cause `etcd` to fall over in the
middle of our test suites. We don't control the installation of `etcd`
so we try to influence storage via system-wide environment.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>