Add shell scripts that underpin the PR approval process
In the next sprint, we will need to programmatically approve or deny
pull requests for each branch of each repository depending on the phase
of the development cycle that we are in as an organization in order to
ensure that breaking changes do not make it into the code that is
deployed to OpenShift Online at the end of every sprint.

In order to do this, each job that handles `[merge]` tags for all of the
separate repositories will call the `approve.sh` script on the Jenkins
master that it is running on. This will call into an approver script via
a symlink that will have been set up by a team lead using the
appropriate Jenkins job previously.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>