Remove `/tmp/openshift` from the artifact list
Origin and the `hack/lib` utilities have moved to place salient data not
at `/tmp/openshift` but under `origin/_output/scripts` and return the
temporary directory to an actually temporary location that should be
used for scratch space. We are inadvertently bringing back many GB of
build artifacts per job if we slurp back `/tmp/openshift`.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>