Run daily extended tests on a cluster installed by OpenShift-Ansible
The current daily extended test jobs run off of an AMI that contains the
pre-build Origin and OpenShift Ansible binaries, RPMs and images. The
tests, however, set up a small OpenShift cluster to run on themselves.
In contrast, the end-to-end tests in the main Origin merge queue run on
top of a cluster installed by OpenShift-Ansible. Switching the parent
job configuration on these daily jobs will cause them to run like the
Origin end-to-end tests do.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>