Add a new 'tooling' severity to the merge approvers
Developers working on the tooling and support code in a repository do
not necessarily have product bugs to fix during dev-cut and/or stage-
cut. Furthermore, their code often has no possible impact on a deploy-
ment or product build, so the restriction for merges is not appropriate
like it is for code that is integrated into the product itself. There-
fore, it was deemed necessary to add a `'tooling'` severity to these
approvers to enable developer productivity.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>