disableConcurrentBuilds()
If two builds of the same version run at the same time (e.g. 3.6 online:int and 3.6 online:stg), image pushing will be affected. This is because the push_image process will find the current version of the image in the dist-git Dockerfile. This is indeterminate if two 3.6 builds are running simultaneously & updating dist-git. 
Once online:stg has a separate dist-git branch, this can be fixed as the race condition will be eliminated.