Adding EARLY_LATEST_HACK
At present, the CI for the Cluster Lifecycle team pulls the :latest image for ose. It expects this version to be found in the latest rhaos puddle. However, there is a long period of time these two versions being consistent because the build pushes a large number of images in the interim.

We have worked around this by only building at night, but recovering from build issues sometimes requires we try builds during the day.

This flag will allow us to hack around this by creating the "latest" link early. This is not a safe operation since other teams may be triggering off the existence of this link and immediately trying to pull images. However, this is the best we can do until we can fix the CI job.