Add smoke test for clusters on upgrade
This adds a new job to run smoke tests on a cluster, and calls that job during
the upgrade job to run a smoke test after the cluster has been upgraded. If
smoke test fails, a noticeable warning is added to the upgrade result's
notification.

Also some improvements to the script to run cluster smoke tests to make it less
prone to false positives.