force logging off for origin install
This goes with https://github.com/openshift/openshift-ansible/pull/5511
When this PR merges, we will need to explicitly set
`openshift_logging_install_logging=True` when installing logging - it
will not use the default in the role anymore.
This also works around the problem in the PR by forcing the origin
install to **not** install logging.