Use wrapper + call cri-o playbooks
It's good practice for testing, to have a static/version-locked execution
environment.  It's also desirable to not "hide" any of this environment
setup outside of the test-content.  Since there's an existing
wrapper-script and requirements file within the cri-o repository,
make use of it to serve these ends.

The result is, the only prerequisites required for executing the
cri-o playbooks (via the wrapper) are boring, bog-standard, and
documented in the playbook README:

    ``python2-virtualenv gcc openssl-devel
      redhat-rpm-config libffi-devel python-devel libselinux-python rsync
      yum-utils python3-pycurl python-simplejson``

Signed-off-by: Chris Evich <cevich@redhat.com>