Add support for changing master_root_volume_size
Previously, upon entering any generated sjb job, the oct tool is already
installed, and a host already provisioned.  It's impossible to make
changes to things like disk-space after generating a job's XML.

Work around this by checking for the presence of an ``$ROOT_VOLUME_SIZE``
env. var when initially installing/configuring oct.  When present,
take it's contents as the value of the master_root_volume_size oct option.

This allows jobs to add ROOT_VOLUME_SIZE to the parameter
section of their job definitions, optionally along with a default value.
This also makes it possible to alter the value for one specific build as
needed.

Signed-off-by: Chris Evich <cevich@redhat.com>