Fix image registry jobs
A parent job of test_branch_image_registry_extended,
origin_release_with_ecosystem.yml, adds openshift,image-registry=master
to sync. But this is the repository that triggers the job, so it causes
the validation added on the PR mentioned above to fail.

This works around the issue by overriding sync and removing
openshift,image-registry=master from it. Not the ideal solution, but it
unblocks the folks from image registry while we work on a better one.