Sync GSSAPI branch with origin master and do not build-extended-test
The sync with origin master guarantees that the branch job and the
pull request job behave in the same way.

We make sure to build the docker registry image to avoid failures in
CI where the docker registry fails to rollout due to an image pull
error.

The call to "make build-extended-test" is unnecessary and thus is
removed to save CI resources.

Signed-off-by: Monis Khan <mkhan@redhat.com>