crio post-actions for gce upload
The 'jq' and 'docker' commands are required for results upload.
Installing them via the playbook is fraught with pitfalls and
complexities. Esp. having docker installed and/or daemon running
while crio tests are executing (they share some config files and
resources).

Doing this as a post-action in the sjb job is not ideal, but has fewer
complexities.  This also makes it easy to trade out docker for podman at a
later date.  Main danger seems to be failures of jq or docker.  Those
will be masked in the job (``trap "exit 0" EXIT``).  Only made obvious
by the lack of GCE results for the run.

Signed-off-by: Chris Evich <cevich@redhat.com>