Abort on image mirror failures. Save input artifacts. Retry mirroring.
Previously this job would catch and hold on to any errors during 'oc
image mirror'. This was nice during initial testing, but now we need
it to actually fail when there is a mirroring problem. However, we
want to give the process a chance to fail so we wrap the mirroring in
a 3x retry block.

Input files to 'oc image mirror' and 'oc apply' are now archived for
easier review later.