logging - use ansible vars.yaml files for branch specific params
This assumes the path for logging code under test is
/data/src/github.com/openshift/origin-aggregated-logging/

The file ci-ansible-common.yaml is for parameters common to
all logging jobs for this branch.
The file ci-ansible-journald.yaml is for `docker --log-driver=journald`
specific configuration.
The file ci-ansible-json-file.yaml is for `docker --log-driver=json-file`
specific configuration.

This requires https://github.com/openshift/origin-aggregated-logging/pull/1530