ocp4: no longer update ose
We no longer want to update the ose repo to track its RPM getting built
in dist-git. Rather than tito tag and then push, the merge is pushed,
and we create/push our own tag on an existing commit. tito tag results
are translated to dist-git but never pushed to ose.

The ose changelog is no longer relevant and so no longer produced.

The logic around determining BUILD_MODE and branches is largely removed
as it's not relevant with our intended versioning scheme.

Later work fixes the fact that we can no longer look in the spec file
for the most recent version.

Builds no longer make ose github releases in 4.y.