ocp4: de-cruft/reorder parameters
A number of parameters were made hardcoded as we have no real need to
vary them. The rest were re-ordered to hopefully be more intuitive.

The logic for determining BUILD_MODE from auto didn't really make sense
any more, and it's unclear whether it will even have any meaning in 4.y.
The only significance that remains is where the RPMs get mirrored to.
Until we can figure out if we care, reduced it to two values with a
likely default.