ocp4: determine version from input or past builds
We're no longer extracting a version from the origin.spec file or
updating it in any way. The user can specify a version for the build, or
reuse/bump the version from the latest build if it's there, or just use
4.y.0 if not.

There were parts of the ose clone automation that try to parse the
origin.spec and determine the previous version/release there, which for
4.y spec files (being almost exactly the upstream spec file now that
tito tags are not pushed) does not even find something it can parse, so a
little bit of parallel logic was required to omit that for 4.y.

A change that got mixed in is that deleting the previous doozer
workspace (which has been known to take some time) is done in parallel
with cloning ose.

Since we no longer tie container versioning to the oc binary, and NVRs
will always be unique, untagging openshift RPM on failure is unnecessary
and that logic has been removed.