buildlib, ocp(4): fix dockerfile reconciliation notices
This fixes a very broken code path. Essentially, reconciliation notices
would only be sent if they were for a special repo that had an alias
passed in from the job. Most repos don't have that, and the real WTF is
that the message simply didn't get sent if it couldn't tell them where
the original source is. And the only sign of this problem is an echo to
the console logs.

With this, what was a "failure to resolve alias" simply results in the
message being sent without trying to point to the upstream Dockerfile.
But if the doozer record.log is working, the failure shouldn't happen
anyway.