refresh-images: fix partial rebuilds
When refreshing only specific images, --ignore-missing-base only works
if the parent image was successfully built at the time the dockerfile
was previously updated. The failure in the case where that's not true is
really confusing. --latest-parent-version will point at the most recent
successful build of the parent which is usually what we want.