Use UTC timestamp as OCP 4 release number
Currently the OCP 4 release number (e.g. `201912130945`) is based on a local timezone. Considering both Red Hat Bugzilla and Errata Tool are using UTC, our life would be easier if the release number uses UTC as well. With UTC timestamp, we can quickly determine if a bugfix is included in a build at first glance.