Sort versions before iterating
When a new release is added, before the job is run to pick up new parameters, Jenkins will not allow you trigger it with the new release value. So make sure we iterate old to new instead of new to old. This should ensure a successful run, which will pick up the new version.