Reflect build/ocp4 result in scheduled builds
As a triage manager, it is confusing to have failed builds when the subsequent
jobs exited with "unstable". It takes some clicks to find out if it is a real
failure or something we can safely ignore (i.e. `freeze_automation`)

I see that `b.result` is being appended to currentBuild's description,
but maybe we could have a better use for that space, such as the reason for failure/unstable.

This PR is more of a conversation starter ... I'd be happy even if we don't merge it,
but end up with a good explanation for why we opted for doing this:
```
currentBuild.result = (b.result == "SUCCESS") ? "SUCCESS" : "FAILURE"
```