rhcos_sync: use whitelist when publishing artifacts to mirror
We don't want customers having direct access to all the artifacts that
RHCOS produces during its build.  (Crafty customers can still figure
it out on their own.)

This provides a whitelisted set of artifacts that we want on
mirror.openshift.com for each GA release.  Unfortunately, this may
change as we produce additional artifacts for future releases or
change policies on what artifacts are on the mirror.