Switch signing-job log archiver to use rclone
Previously we used rsync. Lately we've seen the time require for the
signing job to increase to several hours. Virtually all of that time
is during the log syncing part of the job.

This switches the log archiving to use the 'rclone' tool. This should
speed things up for several reasons:

* No more s3fusefs abstraction layer implementing unnecessary system
  calls against the 'mounted' s3 bucket
* Only considering files created in the last week to copy
* rclone seems very fast in general

The required rclone configuration has been stored in bitwarden