build-sync: compare resourceVersion instead of generation
Checking for image stream updates against the objects generation may
be inaccurate. Instead check against the generation.

https://coreos.slack.com/archives/GDBRP5YJH/p1584643669046500

Justin Pierce  7 minutes ago

The generation number in an imagestream *is* something that might be
slow to update. I think checking the resourceVersion changes will be
immediate & reliable.

I think generation is set by a controller as it works to drive
consistency between desired & actual state. resourceVersion is simpler
and just indicates that a change was made.