Hack'n'Hustle: Auto update pipeline jobs and rebuild images on merge
This adds 2 OpenShift Pipeline jobs `aos-cd-jobs-poll-master` and `update-jobs`.

`aos-cd-jobs-poll-master` is a scheduled job which polls the `master` branch of `aos-cd-jobs` repo
and triggers `update-jobs` job on merge.

`update-jobs` checks changed files from changelog then reconfigures jobs and other API objects defined in `openshift-pipelines/{jobs,infra}` when there is a change.
It also rebuilds the images if Dockerfiles in `openshift-pipelines/images` are changed.

Actual job code is defined in `/vars` with Jenkins Shared Library. The directory is moved to https://github.com/openshift/art-ci-toolkit.git with PR: