olm_bundle: fix up various things
It turns out that splitting an empty string gets you not an empty list,
but rather a list with an empty string as its sole element. So, by
specifying no containers to build, you actually got none, instead of all
as intended. That's fixed.

doozer's debug.log is now archived in case of failure. Also doozer now
uses a working dir instead of /tmp.

Found a few other things that needed tweaks.