ART-2024: Allow promotion of private releases in build/release
The private release controller payloads can be discerned from public payloads since they will include a 'priv' in the nightly name. For example, 4.8.0-0.nightly-s390x-priv-<datestamp> . If an artist enters the name of one of these private release payloads, the release job will:

- Prompt the artist to confirm that embargo dates have lifted
- Build a release payload from the -priv imagestreams.
- Create a new, named (e.g. 4.6.x) release payload for the corresponding PUBLIC release controller.

Also updates the pre-release job to use the new function signature and raise an error if an embargoed nightly is specified.