Signing job fails faster if required files are missing
This changes the order of operations in the signing job. Previously we
would submit a JSON claim signing request and afterwards attempt to
download the matching sha256sum.txt file for a another signing
request.

NOW we check for the sha256sum.txt file FIRST and fail the job if it
is missing. If it exists then operations continue as normal.