Fix 3.Y sweep
After an OCP3 build, the sweep job gets called with the instruction to
attach builds to the advisories. This should not happen, as collecting
builds into advisories should not happen automatically.

This change alters the buildlib.sweep function to assume no builds will
be swept, unless told otherwise.

This also will propagate the DRY_RUN parameter to the sweep job if it
was present in the job calling the sweep job.