pipeline-scripts/release: Do not ping ota-monitor
Vadim has been running an auto-approver for a while now, and with the
recent 1b7b8aeefd (making candidate pr merge directly with hub api
call, 2020-12-09, #2456), it's locally clear that the candidate
pull-request are auto-merging.  This commit adds a note to the message
labeling the auto-merged PRs.  It also removes the monitor ping, since
there's no longer anything that humans need to be aware of at
ART-files-PRs time.