pipeline-scripts/release: Do not ping ota-monitor (#2473)
* pipeline-scripts/release: Do not ping ota-monitor

Vadim has been running an auto-approver for a while now, and with the
recent 1b7b8aeefd (making candidate pr merge directly with hub api
call, 2020-12-09, #2456), it's locally clear that the candidate
pull-request are auto-merging.  This commit adds a note to the message
labeling the auto-merged PRs.  It also removes the monitor ping, since
there's no longer anything that humans need to be aware of at
ART-files-PRs time.

* pipeline-scripts/release: Use lgtm,approved labels to land candidate PRs

The previous merge-API attempt failed:

  + hub api -XPUT repos/openshift/cincinnati-graph-data/pulls/589/merge
  {"message":"2 of 2 required status checks are expected.","documentation_url":"https://docs.github.com/articles/about-protected-branches"}

Instead of trying to merge the pull request, add the labels that Tide
[1] is looking for, and let Tide wait on the presubmit jobs and merge
the pull request when CI passes.

[1]: https://github.com/kubernetes/test-infra/tree/master/prow/tide#tide