Stop syncing 4.2 and 4.3
Alert! This is... not smart. 4.2 and 4.3 need to be synced for ci, so this
commit needs to be reverted soon.

This is to avoid disk space issues on buildvm. A lot of space is
consumed by the artifacts of the resposync job under `/mnt/workspace/reposync/`.
This will be temporarily deleted.