Promote: Run `attach-cve-flaws`
Currently, before promotion, the release manager needs to ensure that
flaw bugs are attached to advisories, and advisories are changed to
RHSAs if needed. If the operator omits this step, the released artifacts
will contain broken links.

This change ensures that, at the time of promotion, the advisories are
changed in kind, foregoing potential operator omissions.

With this change, a leak is introduced. Where *Promote* is only about
the payload, it now operates also on the other advisories. My impression
is that that is acceptable.

This action is not run when:
- this is a dry run
- the advisory is set to -1, manually or otherwise
- this is not a 4stable release