ART-2913 Multi Promote Job (#2606)
https://issues.redhat.com/browse/ART-2913

This is a new wrapper job to trigger multiple promote jobs in parallel for the 3 arches. The usecase is being able to promote all 3 nightlies from one job, something we sorely need :) It doesn't have the advanced params of promote job, in which case ARTist should use the promote job itself. Wanted to keep this pretty basic.