[ART-3329] Extract Cincy PR Slack notification to dedicated function
First step to decouple Slack notification from the creation of PRs.
`sendCincinnatiPRsSlackNotification` is still called inside `openCincinnatiPRs`,
but the goal is to have them completely independent from each other, both called
from `jobs/build/cincinnati-prs/Jenkinsfile`, so it is easier to test them in
isolation, and we'd be able to pass additional arguments to the Slack function
directly, without increasing the amount of arguments passed to `openCincinnatiPRs`.