Perform the changelog computation for every release
console.openshift.com points to ocp-dev-preview/pre-release for dev-preview information. This dev-preview links to ../ocp/candidate which contains whatever is the latest release in a candidate channel. This can be any FC, RC, or GA release until the next release's versions start being added to Cincinnati.  To ensure we always have changelog information in pre-release, we need to compute the changelog for any release, not just pre-GA releases .