jobs/build/cincinnati-prs: Drop CANDIDATE_CHANNEL_ONLY
Since 936c8783f2 (Create only candidate PR, 2021-06-25, #2671), fast
and later promotion has been managed by the graph-data stabilization
robot.  Instead of asking the user to set this property, just inspect
the requested release name to determine if it's a prerelease (which
won't want an errata link) or not (where we'll now return to requiring
an errata link).