Run ocp4 release scans in parallel The original implementation of scan-sources required upstream source for each image to be cloned. This intensive (in terms of CPU and disk space) operations was serialized to prevent exhausting resources.
scan-sources is now relatively quick (5 minutes) and does not clone
upstream source. Even with these improvements, with five releases
being scanned, a change in 4.x can wait to be detected for ~20 minutes
because of the serial scans of releases.

This change allows all scans to happen in parallel - meaning a 4.x
scan is primarily waiting for 4.x builds.