ocp4_scan+rhcos: rebuild when inconsistent
right now, if an rhcos build does not succeed for all arches, it is not
retried and remains inconsistent until the next time ocp4 kicks off an
rhcos build when something else changes. rather than just blindly retry
failed arches (which could delay picking up other changes), we would
like ocp4_scan to kick off a build for all when it sees they are
inconsistent and there's no other reason to run ocp4.

this would not retry if _all_ rhcos builds attempted failed (no
inconsistency), but that seems unlikely to occur enough to be a problem.