ocp4: Build any necessary bundles
In case an ocp4 build gets interrupted after the operator was built, but
before the bundle build completed, the build of the bundle will not be
retried.

This commit changes that behavior, it calls olm_bundle indiscriminately
from the ocp4 job, picking up if a previous bundle was not attempted.
This mimics the approach in pyartcd prepare-release, where the doozer
bundle command gets called for all bundles (and in the general case has
othing to do).

The custom job also calls olm_bundle. The behavior of being selective
has not been changed, as a targeted approach is generally desired
behavior there.