rhcos_sync: use installer data instead of m-os-c
We recently recognized that we really need to be getting the default
RHCOS bootimages according to the installer data file (to match what the
installer downloads by default), not according to machine-os-content
(which may not even have bootimages).

Also fixes up the "normal" case of when we need this (i.e. for public releases):
* Looks at the quay release registry instead of registry.ci
* Fixes a bug with release tag parsing for non-nightly tags