Permit job run which forces image updates
By default, this job will not mirror to upstream if an image already exists upstream. The reason for this is that we want to validate that a new golang builder image is viable before we start pushing it out to CI. The assess this, the `ocp4` job runs mirror without `--only-if-missing` when it successfully builds the OCP api server. 
This option adds the ability to not have to wait for the ocp4 job to successfully run an api server build.