Microshift nightlies: Move to own assembly
When running a test build in assembly `test`, the microshift build will
be found when creating the repository. If this is for a GA or RC
release, the build will get signed. I think this should not be a side
effect of running a test.

Not sure if anything depends on the tooling being in the test assembly.