plashet: Allow previous for rhcos consumed rpms
RHCOS passes a package lock from the x86_64 build on to other arches. If
the ose repo has changed in the mean while, the build will fail.

This change allows two builds from rhcos consumed rpms to be included in
the internal ose repo, so the builds will not choke on that.