Let microshift advisory depend on rpm advisory (#3647)
Following [this
discussion](https://redhat-internal.slack.com/archives/CJARLA942/p1682323495678609),
the microshift advisory should get onboarded to the auto-release release
driver feature. As microshift has a dependency on both images and and
cri-o, letting it depend on the rpm advisory, as the rpm advisory
already depends on the image advisory.