Allow multiple ironic packages versions in plashets
Metal Platform Team needs to pin specific packages for precise
tests and updates.
With this PR we introduce a new list of packages specific for
ironic containers that are allowed to be present in multiple
versions in the plashets.
We'll start covering only the main packages but we may need to
expand the list in the future, separating it fromthe main list
of packages is the most sensible solution.