build-fbc: Use registry config file for pull secret
This is to allow to build FBC for Brew built OLM bundles.

Currently username and password is used for registry auth, but it has a
limitation that `oc` command will use the same login for all registries.
It makes impossible to use Konflux image registry and Brew registry at
the same time. By switching to registry config file, we use different
credentials for different registries.