promote: Remove SKIP_ATTACHED_BUG_CHECK option
Proposing to remove the switch SKIP_ATTACHED_BUG_CHECK. This prevents
this check from running, and think this should always run. If it is
needed to ignore the result, please create a PR to the assembly
definition with content like:

```yaml
promotion_permits:
  - code: ATTACHED_BUGS
    why: |
      - OCPBUGS-123 is fixed in 4.16.56 but OCPBUGS-122 did not make it into 4.17.45
      - OCPBUGS-456 is fixed in 4.16.56 but OCPBUGS-455 did not make it into 4.17.45
```

This enables to query assembly definitions from when this happened. This
info may get queried by ERT or other interested parties.